clean useInfo panel
